Device/IoT-SDK/requirements.txt
azure-iot-device
azure-core
azure.storage.blob